# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank Project API Documentation",
        "description": "API endpoints for managing departments, products, and product types in the Bank Project.",
        "contact": {
            "email": "tvoj.email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Bank Project API Server"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Dohvati sve departmane",
                "description": "Vraca listu svih departmana.",
                "operationId": "1cd271eb5c283ee0a0281bcde5ebc513",
                "responses": {
                    "200": {
                        "description": "Uspjesno dohvacena lista departmana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Department"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup - potrebna Basic HTTP autentifikacija."
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Create a new department",
                "description": "Creates a new department record.",
                "operationId": "44ee53e947eb96a620abf4b108929e32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NAME"
                                ],
                                "properties": {
                                    "NAME": {
                                        "type": "string",
                                        "example": "Finance"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Department created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Department"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "NAME": [
                                                    "The NAME field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/departments/{id}": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Get department by ID",
                "description": "Returns a single department by its ID.",
                "operationId": "3d86649b5467e6456d18072469e417ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Department"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Department not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "put": {
                "tags": [
                    "Departments"
                ],
                "summary": "Update an existing department",
                "description": "Updates an existing department record by ID.",
                "operationId": "06350c81d509a48b638dc5cd0a787e66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NAME"
                                ],
                                "properties": {
                                    "NAME": {
                                        "type": "string",
                                        "example": "Human Resources"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Department"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Department not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "NAME": [
                                                    "The NAME field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            },
            "delete": {
                "tags": [
                    "Departments"
                ],
                "summary": "Delete a department",
                "description": "Deletes a department record by ID.",
                "operationId": "3d279271ee75bb2050d3c864f44ac456",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Department deleted successfully (No Content)"
                    },
                    "404": {
                        "description": "Department not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Department not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup"
                    }
                },
                "security": {
                    "basicAuth": {}
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Department": {
                "title": "Department",
                "description": "Department model",
                "properties": {
                    "DEPT_ID": {
                        "title": "ID",
                        "description": "ID of the department",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "NAME": {
                        "title": "Name",
                        "description": "Name of the department",
                        "type": "string",
                        "example": "IT Department"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Department"
                }
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "HTTP Basic Authentication (Koristi e-mail i lozinku testnog korisnika)"
            }
        }
    },
...
    "security": [
        {
            "basicAuth": []
        }
    ]
}